# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Документация для API",
        "contact": {
            "email": "alekscygankov20@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/password-saver-api/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Регистрация пользователя",
                "operationId": "registerUserPasswordSaverApi",
                "requestBody": {
                    "description": "Данные для регистрации пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Имя пользователя, должно быть уникальным",
                                        "type": "string",
                                        "example": "TestName"
                                    },
                                    "email": {
                                        "description": "Email пользователя, должен быть уникальным",
                                        "type": "string",
                                        "example": "test@gmail.com"
                                    },
                                    "password": {
                                        "description": "Пароль должен содержать минимум 8 символов, включать буквы в верхнем и нижнем регистрах, а также цифры",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Test1234"
                                    },
                                    "password_confirmation": {
                                        "description": "Подтверждение пароля должно совпадать с паролем",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Test1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID зарегистрированного пользователя",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "Имя пользователя",
                                                    "type": "string",
                                                    "example": "TestName"
                                                },
                                                "email": {
                                                    "description": "Email пользователя",
                                                    "type": "string",
                                                    "example": "test@gmail.com"
                                                },
                                                "token": {
                                                    "description": "Bearer токен для авторизации",
                                                    "type": "string",
                                                    "example": "1|0DNmmQtybPzvK2Gp50HnsGUI3JUp2iXVMunfmlb04cc5aa33"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Текст ошибки валидации"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Текст ошибки сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/password-saver-api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Вход пользователя",
                "operationId": "loginUserPasswordSaverApi",
                "requestBody": {
                    "description": "Данные для входа пользователя",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Email пользователя",
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@gmail.com"
                                    },
                                    "password": {
                                        "description": "Пароль пользователя",
                                        "type": "string",
                                        "format": "password",
                                        "example": "Test1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "description": "ID пользователя",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "description": "Имя пользователя",
                                                    "type": "string",
                                                    "example": "TestName"
                                                },
                                                "email": {
                                                    "description": "Email пользователя",
                                                    "type": "string",
                                                    "example": "test@gmail.com"
                                                },
                                                "token": {
                                                    "description": "Токен для авторизации",
                                                    "type": "string",
                                                    "example": "2|1oDxpQzgmAXCHiRzGaBjobeF83wMoqzN77DdThpzbc4846db"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка авторизации"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Текст ошибки валидации"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/password-saver-api/logout": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Выход пользователя",
                "operationId": "logoutUserPasswordSaverApi",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об успешном выходе",
                                            "type": "string",
                                            "example": "Успешно вышел из системы"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка аутентификации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Пользователь не аутентифицирован",
                                            "type": "string",
                                            "example": "Текст ошибки"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Сообщение об ошибке при удалении токена",
                                            "type": "string",
                                            "example": "Ошибка сервера"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}